sync

+++ glazewm/install.ps1
@@ -3,4 +3,6 @@ Write-Host "installing glazewm..."
 Invoke-Expression "winget install glazewm"
 Invoke-Expression "winget uninstall zebar"
 
-Remove-Item -Path "$env:USERPROFILE/.glzr" -Recurse -Force
+if (Test-Path -Path "$env:USERPROFILE/.glzr") {
+    Remove-Item -Path "$env:USERPROFILE/.glzr" -Recurse -Force
+}
